managerd/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.0
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configparser==5.2.0
cryptography==37.0.4
Deprecated==1.2.13
exceptiongroup==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
grpcio==1.51.1
grpcio-tools==1.51.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==2.0.0
IP2Location==8.8.0
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.0.3
keyring==23.6.0
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.0
osquery==3.0.7
packaging==21.3
pandas==1.4.3
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==4.22.1
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.15.0
pygrok==1.0.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.2.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-iptables==1.0.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
regex==2022.6.2
requests==2.28.1
SecretStorage==3.3.2
six==1.16.0
SQLAlchemy==1.4.39
thrift==0.16.0
tomli==2.0.1
urllib3==1.26.9
vine==5.0.0
vt-py==0.14.0
wcwidth==0.2.5
Werkzeug==2.0.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
suricataparser==1.0.0